backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
psycopg2-binary==2.9.9
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
google-cloud-vision==3.4.5
ultralytics==8.2.8
scikit-learn==1.3.2
matplotlib==3.8.2
pandas==2.1.4
# SAM (Segment Anything Model) dependencies
segment-anything==1.0
supervision==0.18.0
transformers==4.36.2

# Depth Anything V2 dependencies
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
huggingface-hub>=0.10.0
transformers>=4.20.0

# Image processing
Pillow>=8.0.0
numpy>=1.21.0
scipy>=1.7.0

# For downloading models
tqdm>=4.60.0 